requirement.txt
SpeechRecognition==3.8.1
pyttsx3==2.90
requests==2.31.0
google-generativeai==0.4.1
duckduckgo-search==4.4
python-dotenv==1.0.1
PyMuPDF==1.23.21
pyaudio==0.2.14